fix: continue parsing beyond element depth cap

+++ packages/js/src/types.ts
@@ -165,8 +165,8 @@ export interface ElementNode extends Node {
   context?: PluginContext
   /** ID of the tag for fast handler lookup */
   tagId?: number
-  /** Map of tag names to their nesting count (using Uint8Array for performance) */
-  depthMap: Uint8Array
+  /** Map of tag names to their nesting count */
+  depthMap: Uint16Array
   /** Plugin outputs collected during processing */
   pluginOutput?: string[]
 }
@@ -227,7 +227,7 @@ export interface Node {
  */
 export interface MdreamProcessingState {
   /** Map of tag names to their current nesting depth - uses TypedArray for performance */
-  depthMap: Uint8Array
+  depthMap: Uint16Array
 
   /** Current overall nesting depth */
   depth: number
